% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeRem.R
\name{largeRem}
\alias{largeRem}
\title{Largest remainder method to round percentages}
\usage{
largeRem(pcVec)
}
\arguments{
\item{pcVec:}{a vector}
}
\description{
This function helps rounding up our down a list of percentages
(that add up to 100) into a list of integers that still add up to 100
}
\examples{
b <- c(20.5,60.3,19.2)
largeRem(b)
[1] 21 60 19
}
